requirements.txt
discord.py == 1.3.3
tweepy == 3.8.0
pyfiglet == 0.8.post1
aiohttp == 3.6.2